requirements.txt
--index-url https://download.pytorch.org/whl/nightly/cu121
torchao-nightly
wandb
schedulefree
transformers
tiktoken